marketplace-stealth-scraper/requirements.txt
python-dotenv
playwright
requests
PyYAML
